--- config.json
@@ -1,1 +1,0 @@
-{"cases_endpt": "https://api.gdc.cancer.gov/cases", "data_endpt": "https://api.gdc.cancer.gov/data/", "primary_site": "breast", "expand": ["diagnoses", "files"], "format": "JSON", "size": "5", "dir": "E:/Przyrod-master/0_master_thesis/data/files/2test", "tumor_stages": {"stage_1": ["Stage I", "Stage IA", "Stage IB"], "stage_4": ["Stage IV"]}, "join_files": "True", "join_method": "merge"}
