+++ fpkmfetcher/src/help.html
@@ -1,0 +1,23 @@
+<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.0//EN" "http://www.w3.org/TR/REC-html40/strict.dtd">
+<html>
+    <head>
+        <meta name="qrichtext" content=1" />
+        <style type="text/css">
+            p, li { white-space: pre-wrap; }
+        </style>
+    </head>
+    <body style="font-family:'MS Shell Dlg 2\'; font-size:8.25pt; font-weight:400; font-style:normal;">
+            <p align="center" style= margin-top:12px; margin-bottom:12px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;"><span style= font-size:16pt;">HELP:</span></p>            <p align=center" style= margin-top:12px; margin-bottom:12px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;"><span style= font-size:8pt; font-weight:600;">FPKM files downloader</span><span style= font-size:8pt;"> </span></p>
+            <p style=" margin-top:12px; margin-bottom:12px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;"><span style=" font-size:8pt;">This program is created in order to download FPKM (Fragments Per Kilobase of transcript per Million mapped reads) files from the NCI's Genomic Data Commons (GDC). More information about GDC and FPKM files you could fined on https://gdc.cancer.gov/</span></p>
+            <p style=" margin-top:12px; margin-bottom:12px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;"><span style=" font-size:8pt;">The program is using The GDC Application Programming Interface (API) to connect to the server and download all necessary files </span></p>
+            <p style=" margin-top:12px; margin-bottom:12px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;"><span style=" font-size:8pt;">   </span></p>
+            <p align="center" style= margin-top:12px; margin-bottom:12px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;"><span style= font-size:8pt; font-weight:600;">Guide</span><span style= font-size:8pt;"> </span></p>
+            <p style=" margin-top:12px; margin-bottom:12px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;"><span style=" font-size:8pt;">In order to download FPKM files you should choose or fill next points:</span></p>
+            <p style=" margin-top:12px; margin-bottom:12px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;"><span style=" font-size:8pt;">1) Choose cancer type (Currently only Breast cancer is available)\n</span></p>
+            <p style=" margin-top:12px; margin-bottom:12px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;"><span style=" font-size:8pt;">2) Write number of files that have to be downloaded of each stage</span></p>
+            <p style=" margin-top:12px; margin-bottom:12px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;"><span style=" font-size:8pt;">3) Choose which stages of cancer have to be downloaded</span></p>
+            <p style=" margin-top:12px; margin-bottom:12px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;"><span style=" font-size:8pt;">4) Choose or fill manualy the directory, where files should be downloaded</span></p>
+            <p style=" margin-top:12px; margin-bottom:12px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;"><span style=" font-size:8pt;">5) Fill checkbox if files have to be joined and choose type of joining: if there are more than 200 files per stage is preferred to choose append method, but it will take much longer</span></p>
+            <p style=" margin-top:12px; margin-bottom:12px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;"><span style=" font-size:8pt;">6) Trigger start button</span></p>
+    </body>
+</html>
