+++ requirements/requirements-all.txt
@@ -1,0 +1,5 @@
+PyQt5>=5.15.7
+requests>=2.28.2
+pandas>=1.5.3
+logmuse>=0.2.7
+coloredlogs>=15.0.1
